04-MINI-PROJECT/DRF-Practice/03_HelloDRF-withIgnore/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
black==22.8.0
click==8.1.3
colorama==0.4.5
Django==3.2.13
django-bootstrap-v5==1.0.11
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
